backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
pytesseract
Pillow
numpy
pandas
scikit-learn
python-dotenv
requests
